requirements.txt
flask
flask-socketio
python-dotenv
soniox
python-socketio[client]
deepgram-sdk
